# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TERA EXAM Sarkari exam platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /syllabus:
    get:
      operationId: listSyllabus
      tags: [syllabus]
      summary: List published syllabus PDFs
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Syllabus entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Syllabus" }
  /syllabus/{id}:
    get:
      operationId: getSyllabus
      tags: [syllabus]
      summary: Get a published syllabus entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Syllabus entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Syllabus" }
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Public: Exam Prep (subjects -> topics -> MCQs)
  # ---------------------------------------------------------------------
...
  /subjects/{id}/topics:
    get:
      operationId: listTopicsBySubject
      tags: [exam-prep]
      summary: List topics for a subject
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Topics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Topic" }
...
  /admin/syllabus:
    get:
      operationId: adminListSyllabus
      tags: [admin]
      summary: List all syllabus entries (any status)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PublishStatus" }
      responses:
        "200":
          description: Syllabus entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Syllabus" }
    post:
      operationId: createSyllabus
      tags: [admin]
      summary: Create a syllabus entry
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyllabusInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Syllabus" }
  /admin/syllabus/{id}:
    patch:
      operationId: updateSyllabus
      tags: [admin]
      summary: Update a syllabus entry (also used for one-click publish via status field)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyllabusUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Syllabus" }
        "404":
          description: Not found
    delete:
      operationId: deleteSyllabus
      tags: [admin]
      summary: Delete a syllabus entry
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Admin: Subjects / Topics / Questions (MCQ bank management)
  # ---------------------------------------------------------------------
  /admin/subjects:
    get:
      operationId: adminListSubjects
      tags: [admin]
      summary: List all subjects
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Subject" }
    post:
      operationId: createSubject
      tags: [admin]
      summary: Create a subject
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subject" }
  /admin/subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [admin]
      summary: Update a subject
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subject" }
        "404":
          description: Not found
    delete:
      operationId: deleteSubject
      tags: [admin]
      summary: Delete a subject
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /admin/topics:
    get:
      operationId: adminListTopics
      tags: [admin]
      summary: List all topics, optionally filtered by subject
      security: [{ bearerAuth: [] }]
      parameters:
        - name: subjectId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Topics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Topic" }
    post:
      operationId: createTopic
      tags: [admin]
      summary: Create a topic
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TopicInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Topic" }
  /admin/topics/{id}:
    patch:
      operationId: updateTopic
      tags: [admin]
      summary: Update a topic
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TopicUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Topic" }
        "404":
          description: Not found
    delete:
      operationId: deleteTopic
      tags: [admin]
      summary: Delete a topic
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublishStatus:
      type: string
      enum: [draft, published]
...
    Syllabus:
      type: object
      required: [id, examName, title, pdfUrl, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        examName: { type: string }
        title: { type: string }
        pdfUrl: { type: string }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SyllabusInput:
      type: object
      required: [examName, title, pdfUrl]
      properties:
        examName: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        pdfUrl: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
    SyllabusUpdate:
      type: object
      properties:
        examName: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        pdfUrl: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }

    # --- Exam prep: Subject / Topic / Question -------------------------
    Subject:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    SubjectInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
    Topic:
      type: object
      required: [id, subjectId, name, slug, createdAt]
      properties:
        id: { type: integer }
        subjectId: { type: integer }
        name: { type: string }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
    TopicInput:
      type: object
      required: [subjectId, name, slug]
      properties:
        subjectId: { type: integer }
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
    TopicUpdate:
      type: object
      properties:
        subjectId: { type: integer }
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
...