# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TERA EXAM Sarkari exam platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  /** @nullable */
  description?: string | null;
}